add ResponseEntity in AccountController

+++ src/main/java/com/example/lmstemplate01/services/AccountService.java
@@ -4,13 +4,14 @@ import com.example.lmstemplate01.configurations.MapperAccount;
 import com.example.lmstemplate01.dto.AccountDTO;
 import com.example.lmstemplate01.model.Account;
 import com.example.lmstemplate01.repositoryJPA.AccountRepository;
+import com.example.lmstemplate01.utils.MLSTemplateRuntimeException;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -21,15 +22,20 @@ public class AccountService implements AccountServiceInterface {
     @Transactional
     @Override
     public AccountDTO createAccount(AccountDTO accountDTO) {
-        Account account = mapperAccount.toAccount(accountDTO);
-        accountRepository.save(account);
-        return mapperAccount.toAccountDTO(account);
+        if (accountRepository.existsByUsername(accountDTO.getUsername())) {
+            throw new MLSTemplateRuntimeException("User with name " + accountDTO.getUsername() + " exists.");
+        } else {
+            Account account = mapperAccount.toAccount(accountDTO);
+            accountRepository.save(account);
+            return mapperAccount.toAccountDTO(account);
+        }
     }
 
     @Transactional
     @Override
     public AccountDTO updateAccount(AccountDTO accountDTO, Long id) {
-        Account account = accountRepository.findById(id).orElseThrow();
+        Account account = accountRepository.findById(id).orElseThrow(
+                () -> new EntityNotFoundException("Account with id " + id + " not found"));
         account.setUsername(accountDTO.getUsername());
         account.setEmail(accountDTO.getEmail());
         account.setPassword(accountDTO.getPassword());//TODO
@@ -41,23 +47,32 @@ public class AccountService implements AccountServiceInterface {
     @Transactional
     @Override
     public void deleteAccount(Long id) {
-        accountRepository.deleteById(id);
+        if (accountRepository.existsById(id)) {
+            accountRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Account doesn't exist.");
+        }
     }
 
     @Transactional(readOnly = true)
     @Override
     public AccountDTO getAccount(Long id) {
-        Account account = accountRepository.findById(id).orElseThrow();
+        Account account = accountRepository.findById(id).orElseThrow(
+                () -> new EntityNotFoundException("Account with id " + id + " not found"));
         return mapperAccount.toAccountDTO(account);
     }
 
     @Transactional(readOnly = true)
     @Override
-    public List<AccountDTO> getAllAccounts() {// Need to retrieve PageRequest?
+    public List<AccountDTO> getAllAccounts() {
         List<Account> users = accountRepository.findAll();
-        List<AccountDTO> usersDTO = new ArrayList<>();
-        users.forEach(account -> usersDTO.add(mapperAccount.toAccountDTO(account)));
-        return usersDTO;
+        if (!users.isEmpty()) {
+            List<AccountDTO> usersDTO = new ArrayList<>();
+            users.forEach(account -> usersDTO.add(mapperAccount.toAccountDTO(account)));
+            return usersDTO;
+        } else {
+            throw new EntityNotFoundException();
+        }
     }
 
     @Transactional(readOnly = true)
